IV курс/Уеб сървърно програмиране/проект/Car-management/requirements.txt
adb-shell==0.4.3
alabaster==0.7.12
asgiref==3.8.1
Babel==2.10.3
beautifulsoup4==4.12.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==38.0.4
distlib==0.3.6
Django==4.2.14
django-freeze==0.11.1
djangorestframework==3.15.2
docutils==0.19
filelock==3.12.2
ghost @ git+https://github.com/EntySec/Ghost@a977a11314c63f2bef7f6ca3e1fe5108c07469e7
idna==3.4
imagesize==1.4.1
importlib-metadata==5.1.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pipenv==2023.7.11
platformdirs==3.8.1
pyasn1==0.4.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
roman==3.3
rsa==4.9
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==1.26.12
virtualenv==20.23.1
virtualenv-clone==0.5.7
xmltodict==0.14.2
zipp==3.10.0
